1. 로깅 설정

+++ blueRibbon/src/main/java/com/blueRibbon/notice/dao/NoticeDao.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.blueRibbon.notice.model.Notice;
 
-public interface NoticeDao extends JpaRepository<Notice, Integer> {
+public interface NoticeDao extends JpaRepository<Notice, Long> {
 
 }
